Avoid int64 to int truncation in WAIT, SLOWLOG, and list indexing
CodeQL flagged several int64 -> int conversions of unbounded
user-supplied values. None are reachable on the amd64 build target,
where int is 64-bit, but the narrowing is unnecessary in each case:

- WAIT: compare the replica count in int64 rather than narrowing the
  parsed argument, removing the conversion entirely.
- SLOWLOG GET: clamp the limit to math.MaxInt before converting, since
  Entries takes an int.
- normalizeListIndex: do the arithmetic in int64 and saturate
  out-of-range indexes to -1 or length. normalizeListRange only tests
  for < 0 and >= length, so behaviour is unchanged.

+++ internal/command/commands.go
@@ -304,7 +304,7 @@ func (e *Executor) handleWait(ctx context.Context, request *Request) (server.Exe
 		"target_offset", targetOffset,
 		"currently_acked", ackedReplicas,
 	)
-	if ackedReplicas >= int(replicas) || timeoutMillis == 0 {
+	if int64(ackedReplicas) >= replicas || timeoutMillis == 0 {
 		return e.finishWaitResult(replicas, ackedReplicas, targetOffset, startedAt, false), nil
 	}
 
@@ -356,7 +356,7 @@ func (e *Executor) waitForReplicaAcknowledgements(ctx context.Context, replicas 
 
 	for {
 		ackedReplicas, notified := e.countReplicasAtOrAboveWithNotify(targetOffset)
-		if ackedReplicas >= int(replicas) {
+		if int64(ackedReplicas) >= replicas {
 			return e.finishWaitResult(replicas, ackedReplicas, targetOffset, startedAt, false), nil
 		}
